create_freshservice_ticket/requirements.txt
boto3==1.24.89
botocore==1.27.89
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
jmespath==1.0.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
